% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cities}
\alias{cities}
\title{U.S. Cities}
\format{
\code{sf} object
\describe{
\item{\code{name}}{City name.}
\item{\code{state}}{City state.}
\item{\code{pop_2020}}{City population in 2020}
\item{\code{GEOID}}{Census GEOID for the corresponding Census Designated Place.}
\item{\code{geometry}}{The \code{sf} geometry column containing the geographic information.}
}
}
\usage{
data("cities")
}
\description{
This data contains the location, name, and 2020 population of U.S. cities and
large towns.
}
\examples{
data(cities)
}
\concept{data}
