% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{dra}
\alias{dra}
\title{Dave's Redistricting Colors}
\format{
character vector of length 5
}
\usage{
dra
}
\description{
Dave's Redistricting App classic color palette.
}
\examples{
dra
}
\concept{colors}
\keyword{datasets}
