% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{fivethirtyeight}
\alias{fivethirtyeight}
\title{FiveThirtyEight Colors}
\format{
character vector of length 5
}
\usage{
fivethirtyeight
}
\description{
Five Thirty Eight's redistricting color scheme for various levels of political
competitiveness.
}
\examples{
fivethirtyeight
}
\concept{colors}
\keyword{datasets}
