% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggredist-package.R
\docType{package}
\name{ggredist-package}
\alias{ggredist}
\alias{ggredist-package}
\title{ggredist: Scales, Palettes, and Extensions of 'ggplot2' for Redistricting}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides `ggplot2` extensions for political mapmaking, including new geometries, easy label generation and placement, automatic map coloring, and map scales, palettes, and themes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alarm-redist/ggredist}
  \item \url{https://alarm-redist.github.io/ggredist/}
  \item Report bugs at \url{https://github.com/alarm-redist/ggredist/issues}
}

}
\author{
\strong{Maintainer}: Christopher T. Kenny \email{christopherkenny@fas.harvard.edu}

Authors:
\itemize{
  \item Cory McCartan \email{cmccartan@g.harvard.edu}
}

}
\keyword{internal}
