% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{ggredist}
\alias{ggredist}
\title{Redistricting Color Palettes}
\format{
list of character vectors of type \code{palette} (which supports a
\code{plot()} generic for visualization)
}
\usage{
ggredist
}
\description{
Included palettes:
\itemize{
\item \strong{partisan}, a perceptually uniform red-white-blue scale
\item \strong{dra}, the Dave's Redistricting App classic color palette
\item \strong{penn82}, historic categorical color palette from the 1982 published
Pennsylvania congressional district map
\item \strong{randmcnally} and \strong{natgeo}, colors taken from Rand-McNally and National
Geographic political maps
\item \strong{coast} and \strong{larch}, inspired by natural scenery
}
}
\examples{
scales::show_col(ggredist$partisan)
scales::show_col(ggredist$dra)
scales::show_col(ggredist$penn82)
scales::show_col(ggredist$randmcnally)
scales::show_col(ggredist$natgeo)
scales::show_col(ggredist$coast)
scales::show_col(ggredist$larch)

}
\concept{colors}
\keyword{datasets}
