% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{GOP_DEM}
\alias{GOP_DEM}
\title{Party Colors}
\format{
character vector of length 15
}
\usage{
GOP_DEM
}
\description{
Vector of evenly spaced party colors from red to white to blue.
}
\examples{
GOP_DEM

}
\keyword{datasets}
