% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{interstates}
\alias{interstates}
\title{Insterstate Shapefile}
\format{
\code{sf} object
\describe{
\item{\code{name}}{Census Bureau name for the interstate}
\item{\code{geometry}}{The \code{sf} geometry column containing the geographic information.}
}
}
\usage{
data("interstates")
}
\description{
This data contains geometry information for the U.S. Interstate Highway System.
It was processed from the U.S. Census Bureau TIGER/Line Shapefile system.
}
\examples{
data(interstates)
}
\concept{data}
