% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{percent_no_dec}
\alias{percent_no_dec}
\title{Label Percentages with No Decimal}
\usage{
percent_no_dec(...)
}
\arguments{
\item{...}{additional arguments to pass to \code{scales::label_percent}}
}
\value{
returns a labelling function
}
\description{
Label Percentages with No Decimal
}
\examples{
# TODO

}
