% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/538.R
\name{scale_fill_538}
\alias{scale_fill_538}
\alias{scale_color_538}
\title{FiveThirtyEight scales for \code{ggplot2}}
\usage{
scale_fill_538(...)

scale_color_538(...)
}
\arguments{
\item{...}{additional arguments to \verb{ggplot::scale_*} functions}
}
\value{
ggplot scale function
}
\description{
FiveThirtyEight scales for \code{ggplot2}
}
\examples{
scale_fill_538()
scale_color_538()
}
\concept{colors}
