% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alaska.R
\name{scale_fill_alaska}
\alias{scale_fill_alaska}
\alias{scale_color_alaska}
\alias{scale_colour_alaska}
\title{Alaska Color Scales for \code{ggplot2}}
\usage{
scale_fill_alaska(...)

scale_color_alaska(...)

scale_colour_alaska(...)
}
\arguments{
\item{...}{Arguments passed on to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}}
}
\value{
ggplot scale function
}
\description{
Alaska Color Scales for \code{ggplot2}
}
\examples{
library(ggplot2)
data(oregon)

ggplot(oregon, aes(group = county)) +
    geom_district() +
    scale_fill_alaska() +
    theme_map()

}
\concept{colors}
