% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dra.R
\name{scale_fill_dra}
\alias{scale_fill_dra}
\title{Dave's Redistricting App classic scale for \code{ggplot2}}
\usage{
scale_fill_dra(...)
}
\arguments{
\item{...}{Arguments passed on to [ggplot2::discrete_scale()}
}
\description{
Dave's Redistricting App classic scale for \code{ggplot2}
}
\examples{
scale_fill_dra()
}
\concept{scales}
