% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/florida.R
\name{scale_fill_jacksonville}
\alias{scale_fill_jacksonville}
\alias{scale_color_jacksonville}
\alias{scale_colour_jacksonville}
\alias{scale_fill_florida}
\alias{scale_color_florida}
\alias{scale_colour_florida}
\title{Jacksonville and Florida Color Scales for \code{ggplot2}}
\usage{
scale_fill_jacksonville(...)

scale_color_jacksonville(...)

scale_colour_jacksonville(...)

scale_fill_florida(...)

scale_color_florida(...)

scale_colour_florida(...)
}
\arguments{
\item{...}{Arguments passed on to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}}
}
\value{
ggplot scale function
}
\description{
Jacksonville and Florida Color Scales for \code{ggplot2}
}
\examples{
library(ggplot2)
data(oregon)

ggplot(oregon, aes(group = county)) +
    geom_district() +
    scale_fill_jacksonville() +
    theme_map()

ggplot(oregon, aes(group = county)) +
    geom_district() +
    scale_fill_florida() +
    theme_map()

}
\concept{colors}
