% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polimap.R
\name{scale_fill_randmcnally}
\alias{scale_fill_randmcnally}
\alias{scale_color_randmcnally}
\alias{scale_colour_randmcnally}
\alias{scale_fill_natgeo}
\alias{scale_color_natgeo}
\alias{scale_colour_natgeo}
\title{Rand-McNally and National Geographic Color Scales for \code{ggplot2}}
\usage{
scale_fill_randmcnally(...)

scale_color_randmcnally(...)

scale_colour_randmcnally(...)

scale_fill_natgeo(...)

scale_color_natgeo(...)

scale_colour_natgeo(...)
}
\arguments{
\item{...}{Arguments passed on to [ggplot2::discrete_scale()}
}
\description{
Rand-McNally and National Geographic Color Scales for \code{ggplot2}
}
\examples{
library(ggplot2)
data(oregon)

ggplot(oregon, aes(fill = factor(cd_2020))) +
    geom_sf(size = 0) +
    scale_fill_randmcnally() +
    theme_map()

ggplot(oregon, aes(fill = factor(cd_2020))) +
    geom_sf(size = 0) +
    scale_fill_natgeo() +
    theme_map()

}
\concept{colors}
