% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texas.R
\name{scale_fill_texas}
\alias{scale_fill_texas}
\alias{scale_color_texas}
\alias{scale_colour_texas}
\title{Texas Color Scales for \code{ggplot2}}
\usage{
scale_fill_texas(...)

scale_color_texas(...)

scale_colour_texas(...)
}
\arguments{
\item{...}{Arguments passed on to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}}
}
\value{
ggplot scale function
}
\description{
Texas Color Scales for \code{ggplot2}
}
\examples{
library(ggplot2)
data(oregon)

ggplot(oregon, aes(group = county)) +
    geom_district() +
    scale_fill_texas() +
    theme_map()

}
\concept{colors}
