% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/washington.R
\name{scale_fill_washington}
\alias{scale_fill_washington}
\alias{scale_color_washington}
\alias{scale_colour_washington}
\title{Washington Redistricting Commission Color Scales for \code{ggplot2}}
\usage{
scale_fill_washington(...)

scale_color_washington(...)

scale_colour_washington(...)
}
\arguments{
\item{...}{Arguments passed on to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}}
}
\value{
ggplot scale function
}
\description{
Washington Redistricting Commission Color Scales for \code{ggplot2}
}
\examples{
library(ggplot2)
data(oregon)

ggplot(oregon, aes(group = county)) +
    geom_district() +
    scale_fill_washington() +
    theme_map()

}
\concept{colors}
