% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki.R
\name{scale_fill_wiki_rep}
\alias{scale_fill_wiki_rep}
\alias{scale_color_wiki_rep}
\alias{scale_fill_wiki_dem}
\alias{scale_color_wiki_dem}
\alias{scale_fill_wiki_rep_pres}
\alias{scale_color_wiki_rep_pres}
\alias{scale_fill_wiki_dem_pres}
\alias{scale_color_wiki_dem_pres}
\title{Wikipedia Style Election Map scales for \code{ggplot2}}
\usage{
scale_fill_wiki_rep(...)

scale_color_wiki_rep(...)

scale_fill_wiki_dem(...)

scale_color_wiki_dem(...)

scale_fill_wiki_rep_pres(...)

scale_color_wiki_rep_pres(...)

scale_fill_wiki_dem_pres(...)

scale_color_wiki_dem_pres(...)
}
\arguments{
\item{...}{additional arguments to \verb{ggplot::scale_*} functions}
}
\value{
ggplot scale function
}
\description{
Wikipedia Style Election Map scales for \code{ggplot2}
}
\examples{
library(ggplot2)
data(oregon)

ggplot(oregon, aes(fill = ndv / (ndv + nrv))) +
    geom_sf(size = 0) +
    scale_fill_wiki_dem(name = 'Dem Share') +
    theme_map()
}
\concept{colors}
