% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa.R
\name{scale_fill_ye_olde_pa}
\alias{scale_fill_ye_olde_pa}
\title{Ye Olde Pennsylvania scales for \code{ggplot2}}
\usage{
scale_fill_ye_olde_pa(...)
}
\arguments{
\item{...}{Arguments passed on to [ggplot2::discrete_scale()}
}
\description{
Ye Olde Pennsylvania scales for \code{ggplot2}
}
\examples{
# TODO
}
\concept{scales}
