% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_map}
\alias{theme_map}
\title{Map Theme}
\usage{
theme_map(...)
}
\arguments{
\item{...}{additional parameters}
}
\value{
ggplot2 theme
}
\description{
Theme for maps which uses the \code{'Times'} family and has a transparent background.
}
\examples{
library(ggplot2)
data(oregon)

ggplot(oregon, aes(group = county)) +
    geom_district() +
    scale_fill_penn82() +
    theme_map()

}
\concept{theme}
