% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_r21}
\alias{theme_r21}
\title{Plot Theme from Redistricting 2021 Blog}
\usage{
theme_r21(...)
}
\arguments{
\item{...}{}
}
\value{
ggplot2 theme
}
\description{
Plot Theme from Redistricting 2021 Blog
}
