% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{ye_olde_pa}
\alias{ye_olde_pa}
\title{Ye Olde Pennsylvania}
\format{
An object of class \code{character} of length 6.
}
\usage{
ye_olde_pa
}
\description{
Historic categorical color palette for Pennsylvania redistricting.
}
\examples{
ye_olde_pa

}
\concept{colors}
\keyword{datasets}
