% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cities}
\alias{cities}
\title{U.S. Cities}
\format{
\code{sf} object
\describe{
\item{\code{name}}{City name.}
\item{\code{state}}{City state.}
\item{\code{pop_2006}}{City population in 2006.}
\item{\code{capital}}{TRUE if the city is a capital.}
\item{\code{geometry}}{The \code{sf} geometry column containing the geographic information.}
}
}
\usage{
data("cities")
}
\description{
This data contains the location, name, and 2006 population of U.S. cities and
large towns.
}
\examples{
data(cities)
}
\concept{data}
