% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/538.R
\name{scale_fill_538}
\alias{scale_fill_538}
\title{FiveThirtyEight scales for \code{ggplot2}}
\usage{
scale_fill_538(...)
}
\arguments{
\item{...}{additional arguments to \verb{ggplot::scale_*} functions}
}
\description{
FiveThirtyEight scales for \code{ggplot2}
}
\examples{
# TODO
}
\concept{scales}
