% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dra.R
\name{scale_fill_dra}
\alias{scale_fill_dra}
\alias{scale_color_dra}
\alias{scale_colour_dra}
\title{Dave's Redistricting App classic scale for \code{ggplot2}}
\usage{
scale_fill_dra(...)

scale_color_dra(...)

scale_colour_dra(...)
}
\arguments{
\item{...}{Arguments passed on to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}}
}
\value{
ggplot scale function
}
\description{
Dave's Redistricting App classic scale for \code{ggplot2}
}
\examples{
library(ggplot2)
data(oregon)

ggplot(oregon, aes(group = county, fill=county)) +
    geom_district() +
    scale_fill_dra() +
    theme_map()

}
\concept{colors}
