% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/538.R, R/party.R
\name{scale_color_538}
\alias{scale_color_538}
\alias{scale_fill_party_c}
\alias{scale_fill_party_d}
\alias{scale_color_party_c}
\alias{scale_color_party_d}
\alias{scale_color_party_b}
\title{Party scales for \code{ggplot2}}
\usage{
scale_color_538(...)

scale_fill_party_c(
  name = "Democratic share",
  midpoint = 0.5,
  limits = 0:1,
  labels = scales::percent,
  oob = scales::squish,
  ...,
  reverse = FALSE
)

scale_fill_party_d(...)

scale_color_party_c(
  name = "Democratic share",
  midpoint = 0.5,
  limits = 0:1,
  labels = scales::percent,
  oob = scales::squish,
  ...
)

scale_color_party_d(...)

scale_color_party_b(...)
}
\arguments{
\item{...}{additional arguments to \verb{ggplot::scale_*} functions}

\item{name}{name for scale. Default is \code{"Democratic share"}.}

\item{midpoint}{Scale midpoint value. Default is \code{0.5}.}

\item{limits}{Lower and upper limits for scale. Default is \code{0:1}.}

\item{labels}{function to adjust scale labels. Default is \code{scales::percent}.}

\item{oob}{function to deal with out of bounds. Default is \code{scales::squish()}.}
}
\description{
Party scales for \code{ggplot2}
}
\examples{
# TODO
}
\concept{scales}
