% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_district_text.R
\docType{data}
\name{geom_district_text}
\alias{geom_district_text}
\alias{geom_district_label}
\alias{stat_district_coordinates}
\alias{StatDistrictCoordinates}
\alias{GeomDistrictText}
\title{Label Map Regions}
\usage{
geom_district_text(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  adjust = 1,
  check_overlap = FALSE,
  parse = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)

geom_district_label(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  label.padding = unit(0.25, "lines"),
  label.r = unit(0.15, "lines"),
  label.size = 0.25,
  parse = FALSE,
  adjust = 1,
  check_overlap = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)

stat_district_coordinates(
  mapping = NULL,
  data = NULL,
  geom = "text",
  position = "identity",
  na.rm = FALSE,
  adjust = 1,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}}}

\item{data}{The data to be displayed in this layer}

\item{position}{Position adjustment}

\item{na.rm}{if \code{TRUE}, will silently remove missing values from calculations}

\item{show.legend}{Should this layer be included in the legends?}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather than
combining with them.}

\item{...}{Passed onto the underlying geoms.}

\item{geom}{The geometric object to use display the data}
}
\value{
a \code{ggproto} object
}
\description{
Aggregates shapefile according to the \code{group} aesthetic and positions labels
for each region defined by \code{group}.
}
\examples{
library(ggplot2)
data(oregon)

ggplot(oregon, aes(group=county)) +
    geom_district() +
    geom_district_text() +
    scale_fill_randmcnally() +
    theme_map()

ggplot(oregon, aes(group=cd_2020)) +
  geom_district(lwd=0.9, color="#442222") +
  geom_district(aes(group=county), lwd=0.4, lty="dashed", fill=NA) +
  geom_district_text(aes(group=county, label=toupper(county)),
                     size=2.2, check_overlap=T) +
  geom_district_text() +
  scale_fill_penn82() +
  theme_map()

}
\concept{geoms}
\keyword{datasets}
